+++ packages/compiler/package.json
@@ -49,6 +49,10 @@
     "./experimental/typekit": {
       "types": "./dist/src/experimental/typekit/index.d.ts",
       "default": "./dist/src/experimental/typekit/index.js"
+    },
+    "./internals": {
+      "types": "./dist/src/internals/index.d.ts",
+      "import": "./dist/src/internals/index.js"
     }
   },
   "browser": {
