+++ packages/compiler/src/config/config-to-options.ts
@@ -3,7 +3,8 @@ import { createDiagnostic } from "../core/messages.js";
 import { CompilerOptions } from "../core/options.js";
 import { getDirectoryPath, normalizePath } from "../core/path-utils.js";
 import { CompilerHost, Diagnostic, NoTarget } from "../core/types.js";
-import { deepClone, doIO, omitUndefined } from "../utils/misc.js";
+import { doIO } from "../utils/io.js";
+import { deepClone, omitUndefined } from "../utils/misc.js";
 import { expandConfigVariables } from "./config-interpolation.js";
 import { loadTypeSpecConfigForPath, validateConfigPathsAbsolute } from "./config-loader.js";
 import { EmitterOptions, TypeSpecConfig } from "./types.js";
