+++ packages/compiler/src/core/cli/cli.ts
@@ -8,7 +8,7 @@ try {
 
 import yargs from "yargs";
 import { installTypeSpecDependencies } from "../../install/install.js";
-import { typespecVersion } from "../../utils/misc.js";
+import { typespecVersion } from "../../manifest.js";
 import { getTypeSpecEngine } from "../engine.js";
 import { compileAction } from "./actions/compile/compile.js";
 import { formatAction } from "./actions/format.js";
