+++ packages/compiler/src/init/core-templates.ts
@@ -1,6 +1,6 @@
 import { CompilerPackageRoot } from "../core/node-host.js";
 import { resolvePath } from "../core/path-utils.js";
-import { CompilerHost } from "../index.js";
+import type { SystemHost } from "../index.js";
 
 export const templatesDir = resolvePath(CompilerPackageRoot, "templates");
 export interface LoadedCoreTemplates {
@@ -9,7 +9,7 @@ export interface LoadedCoreTemplates {
 }
 
 let typeSpecCoreTemplates: LoadedCoreTemplates | undefined;
-export async function getTypeSpecCoreTemplates(host: CompilerHost): Promise<LoadedCoreTemplates> {
+export async function getTypeSpecCoreTemplates(host: SystemHost): Promise<LoadedCoreTemplates> {
   if (typeSpecCoreTemplates === undefined) {
     const file = await host.readFile(resolvePath(templatesDir, "scaffolding.json"));
     const content = JSON.parse(file.text);
