+++ packages/compiler/src/server/completion.ts
@@ -32,7 +32,8 @@ import {
   TypeSpecScriptNode,
 } from "../core/types.js";
 import { PackageJson } from "../types/package-json.js";
-import { findProjectRoot, loadFile, resolveTspMain } from "../utils/misc.js";
+import { findProjectRoot, loadFile } from "../utils/io.js";
+import { resolveTspMain } from "../utils/misc.js";
 import { getSymbolDetails } from "./type-details.js";
 
 export type CompletionContext = {
